smartknob-connection2/requirements.txt
# SmartKnob Connection Library - Clean Implementation
# Core dependencies for COBS+CRC32+Protobuf protocol

# Serial communication
pyserial>=3.5

# Protocol encoding
cobs>=1.2.0

# Protobuf support
protobuf>=4.0.0

# Protobuf compilation (includes protoc as Python module)
grpcio-tools>=1.50.0

# Development and testing
pytest>=7.0.0
